detalles1

+++ libft/ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: gpaez-ga <gpaez-ga@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/11 17:43:07 by gpaez-ga          #+#    #+#             */
-/*   Updated: 2023/09/11 17:49:30 by gpaez-ga         ###   ########.fr       */
+/*   Updated: 2023/09/16 19:59:32 by gpaez-ga         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strlen(const char *str)
+size_t	ft_strlen(const char *str)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i])
